updated .gitignore to (hopefully) ignore files generated by the build system.
will probably need to stay on top of this better unless there's another better
way of doing it. As I add features, there may be more build files generated
at build time that I don't know about. TODO: find out if there is any way to
completely remove a file from the git repo's history so it doesn't need to be
downloaded: for example, a really big texture file that is not needed.