now using gluLookAt() for camera transformation.
I'm really not all that interested in getting bleeding edge performance
or anything at this time. It just needs to render correctly. The emphasis
should be on simple easy-to-read code, and extensibility.
If anyone wants to revise the camera code and whatnot, give it a try.
what I REALLY need is a sanity check on the physics. I guess I'd better
work on exporting stuff to a scripting interface...